simulator/bridge_tester/requirements_dev.txt
pip==19.0.3
#wheel==0.33.1
pyserial==3.5
